gen.jl Documentation

gen.QuantileVal

gen.QuantileVal — Method

QuantileVal(distribution, quant)

TEXT

Takes distribution type and the quantile (i.e. 50 for 50th quantile) and returns pdf evaluated at that quantile. Is used to calculate joint probalities (we assume independence of types). Example:

julia> using Distributions

julia> include("/Users/AnR/Dropbox/Basic-Profile/basic-program-share/src/solver/modules/gen/src/gen.jl")
Main.##ex-#253.gen

julia> distribution="logistic"; quant = 40
40

julia> gen.QuantileVal(distribution,quant)
0.24

gen.ProblemGenerator

gen.ProblemGenerator — Method

ProblemGenerator(util,shock,cap,shock_distribution,mode)

TEXT

Takes DataFrame inputs for utilities, shocks, capacities, and a shock distribution. Can be switched into DataFrame mode (mode="DF") and CSV mode (mode="CSV"), depending on input type.

Generates the number of types, means, the vector of capacities, the vector giving the distribution of types, and the set of types in three forms:

an array of types by quantile (each row is a type)

an unnormalized array of "raw" types (mean + shock) (each row is a type)

and a vector of types in vector form.

shock_distribution currently supports only "logistic" and "uniform".

Example:

julia> using DataFrames

julia> include("/Users/AnR/Dropbox/Basic-Profile/basic-program-share/src/solver/modules/gen/src/gen.jl")
Main.##ex-#254.gen

julia> cap_df = DataFrame(cap =[.7,.3]); util_df = DataFrame(mean =[15,20]); shock_df = DataFrame(shock=[.1,.5])
2×1 DataFrames.DataFrame
│ Row │ shock   │
│     │ Float64 │
├─────┼─────────┤
│ 1   │ 0.1     │
│ 2   │ 0.5     │

julia> gen.ProblemGenerator(util_df,shock_df,cap_df,"logistic","DF")[1:2]
(4, 2)

julia> gen.ProblemGenerator(util_df,shock_df,cap_df,"logistic","DF")[3]
1×2 Array{Float64,2}:
 0.7  0.3

julia> gen.ProblemGenerator(util_df,shock_df,cap_df,"logistic","DF")[4]
4×2 Array{Float64,2}:
 12.8028  17.8028
 12.8028  20.0
 15.0     17.8028
 15.0     20.0

julia> gen.ProblemGenerator(util_df,shock_df,cap_df,"logistic","DF")[5]
4×1 Array{Any,2}:
 [0.1, 0.1]
 [0.1, 0.5]
 [0.5, 0.1]
 [0.5, 0.5]

julia> gen.ProblemGenerator(util_df,shock_df,cap_df,"logistic","DF")[6]
4×2 Array{Float64,2}:
 0.1  0.1
 0.1  0.5
 0.5  0.1
 0.5  0.5

julia> gen.ProblemGenerator(util_df,shock_df,cap_df,"logistic","DF")[7]
4-element Array{Float64,1}:
 0.07006920415224915
 0.19463667820069208
 0.19463667820069208
 0.5406574394463667