solve.jl Documentation

solve.SolverGLPK

solve.SolverGLPK — Method

SolverGLPK(num_types,num_objects,type_arr,type_probs,cap_vec,rtype_arr,folder,type_vec_shock,print_bool,infocon_bool,eff_bool)

Applies ModelSolver with GLPK Optimizer

julia> using DataFrames

julia> include("/Users/AnR/Dropbox/Basic-Profile/basic-program-share/src/solver/modules/gen/src/gen.jl")
Main.##ex-#255.gen

julia> cap_df = DataFrame(cap =[.3,.7,.5]); util_df = DataFrame(mean =[10,15,20]); shock_df = DataFrame(shock=[.1,.9])
2×1 DataFrames.DataFrame
│ Row │ shock   │
│     │ Float64 │
├─────┼─────────┤
│ 1   │ 0.1     │
│ 2   │ 0.9     │

julia> input = gen.ProblemGenerator(util_df,shock_df,cap_df,"logistic","DF")
(8, 3, [0.3 0.7 0.5], [7.802775422663781 12.802775422663782 17.80277542266378; 7.802775422663781 12.802775422663782 22.19722457733622; … ; 12.19722457733622 17.19722457733622 17.80277542266378; 12.19722457733622 17.19722457733622 22.19722457733622], Any[[0.1, 0.1, 0.1]; [0.1, 0.1, 0.9]; … ; [0.9, 0.9, 0.1]; [0.9, 0.9, 0.9]], [0.1 0.1 0.1; 0.1 0.1 0.9; … ; 0.9 0.9 0.1; 0.9 0.9 0.9], [0.12500000000000008, 0.12500000000000003, 0.12500000000000003, 0.12499999999999996, 0.12500000000000003, 0.12499999999999996, 0.12499999999999997, 0.12499999999999992])

julia> include("/Users/AnR/Dropbox/Basic-Profile/basic-program-share/src/solver/modules/solve/src/solve.jl")
Main.##ex-#255.solve

julia> using JuMP, GLPK, MathOptInterface
WARNING: using JuMP.solve in module ##ex-#255 conflicts with an existing identifier.

julia> const MOI = MathOptInterface
MathOptInterface

julia> import JuMP: GenericAffExpr

julia> output = solve.SolverGLPK(input[1],input[2],input[4],input[7],input[3],input[6],"output",input[5],0,1,0)[1]
8×3 Array{Float64,2}:
 0.0       0.318001  0.658801
 0.0       0.318001  0.658801
 0.0       1.0       0.0
 0.0       0.318001  0.658801
 0.102985  0.215016  0.681999
 0.102985  0.215016  0.681999
 0.0       1.0       0.0
 0.0       0.318001  0.658801

solve.ModelSolver

solve.ModelSolver — Method

ModelSolver(num_types,num_objects,type_arr,type_probs,cap_arr,rtype_arr,m,folder,type_vec_shock,print_bool,infocon_bool,eff_bool)

Given a model m, ModelSolver solves the optimization problem subject to the appropriate constrains.

Given the number of types, number of means, capacity array, type arrays and type distribution, finds the optimal allocation given the following constraints:

Capacity

Feasibility

Non-negativity

(Incentive Compatibility, applied when infocon_bool==1)

(Efficiency, applied when eff_bool==1)

Saves the following tuple when print_bool ==1: (assignment array,type array,type distribution array, vector of raw type vectors, incentive constraint designation). Otherwise, returns (assignment array,type array,type distribution array, vector of raw type vectors). Assignment array takes the form such that element (i,j) gives the probability that type i is assigned to each object j.

Is parametrized such that the type distribution represents unit capacity.