gen.jl Documentation

gen.ProblemGeneratorSimple

gen.ProblemGeneratorSimple — Method

ProblemGeneratorSimple(util,shock,cap,shock_distribution,mode)

Takes input parameters for base utilities, quantiles at which preferences are obtained, capacities of objects, and distribution (given as Distributions.jl type).

Takes DataFrame inputs for utilities, shocks, capacities, and takes distributions for the shocks over each mean. Supports either a single distribution type for all inputs (e.g. Logistic()) or an array of distributions types of size equal to the number of means (e.g. [Logistic(),Logistic(),Logistic()] ). Can be switched into DataFrame mode (mode="DF") and CSV mode (mode="CSV"), depending on input type.

Returns an array of types (where entry (i,j) represents the value type i assigns to object j), an array of probabilities over these types (the distribution), and the capacity array.

Note

shock_distribution currently supports only continuous distributions included in Distributions.jl. For example, Logistic() is a valid input, but logistic is not. Please see this link for more details.

Let us consider an example where the problem specification is given as a set of DataFrames through the command line:

julia> using DataFrames, Distributions

julia> include(abspath("../../../../basic-program-share/src/solver/modules/gen/src/gen.jl"))
Main.##ex-#253.gen

julia> cap_df = DataFrame(cap =[.7,.3]); util_df = DataFrame(mean =[15,20]); shock_df = DataFrame(shock=[.1,.8])
2×1 DataFrames.DataFrame
│ Row │ shock   │
│     │ Float64 │
├─────┼─────────┤
│ 1   │ 0.1     │
│ 2   │ 0.8     │

julia> gen.ProblemGeneratorSimple(util_df,shock_df,cap_df,Logistic(),"DF")[1]
4×2 Array{Float64,2}:
 12.8028  17.8028
 12.8028  21.3863
 16.3863  17.8028
 16.3863  21.3863

julia> gen.ProblemGeneratorSimple(util_df,shock_df,cap_df,Logistic(),"DF")[2]
4-element Array{Float64,1}:
 0.012170296519395718
 0.09814877037022095
 0.09814877037022095
 0.7915321627401624

julia> gen.ProblemGeneratorSimple(util_df,shock_df,cap_df,Logistic(),"DF")[3]
1×2 Array{Float64,2}:
 0.7  0.3

Now, let us consider an example where the problem specification is given as a set of paths to the appropriate CSV files:

julia> using DataFrames, CSV, Distributions

julia> include(abspath("../../../../basic-program-share/src/solver/modules/gen/src/gen.jl"))
Main.##ex-#254.gen

julia> cap_df = DataFrame(cap =[.7,.3]); util_df = DataFrame(mean =[15,20]); shock_df = DataFrame(shock=[.1,.8])
2×1 DataFrames.DataFrame
│ Row │ shock   │
│     │ Float64 │
├─────┼─────────┤
│ 1   │ 0.1     │
│ 2   │ 0.8     │

julia> CSV.write("cap.csv", cap_df, writeheader=true);CSV.write("util.csv", util_df, writeheader=true);CSV.write("shock.csv", shock_df, writeheader=true)
"shock.csv"

julia> gen.ProblemGeneratorSimple("util.csv","shock.csv","cap.csv",[Logistic(),Uniform()],"CSV")[1]
4×2 Array{Float64,2}:
 12.8028  20.1
 12.8028  20.8
 16.3863  20.1
 16.3863  20.8

julia> gen.ProblemGeneratorSimple("util.csv","shock.csv","cap.csv",[Logistic(),Uniform()],"CSV")[2]
4-element Array{Float64,1}:
 0.02185845168520239
 0.08846061520441428
 0.17628002338976875
 0.7134009097206145

julia> gen.ProblemGeneratorSimple("util.csv","shock.csv","cap.csv",[Logistic(),Uniform()],"CSV")[3]
1×2 Array{Float64,2}:
 0.7  0.3

julia> rm("util.csv");rm("shock.csv");rm("cap.csv")

We recommend this second method for ease of use. Please see DataFrames.jl for methods of converting arrays and other data types into the DataFrames type, and CSV.jl for saving data structures as arrays.