solve.jl Documentation

solve.SolverL

solve.SolverL — Method

SolverL(type_arr,type_probs,cap_arr;infocon_bool=true,eff_bool=false)

Applies ModelSolver with GLPK Optimizer. Use for Linear Programming (without prices)

This solver should be used for linear programming problems, as the GLPK optimizer used only accepts such problems. For non-linear problems (using prices), please see SolverNL.

Given the type array, type distribution, and capacity array, finds the optimal allocation given the following constraints:

Capacity: $\Sigma_{\theta \in \Theta} f_{\theta} m_{\theta a} \leq c_a, \forall a \in A$

Feasibility: $\Sigma_{a \in A} m_{\theta a} \leq 1, \forall \theta \in \Theta$

Non-negativity: $m_{\theta a} \geq 0, \forall (\theta,a) \in \Theta \times A$

(Incentive Compatibility, applied when infocon_bool==1): $\Sigma_{a \in A} u_{\theta a} m_{\theta a} - \Sigma_{a \in A} u_{\theta a} m_{\theta' a} \geq 0, \forall \theta,\theta' \in \Theta$

(Efficiency, applied when eff_bool==1): pending

Saves the following tuple when print_bool ==1 to the folder given by folder ==path/to/folder: (assignment array,type array,type distribution array, incentive constraint designation). Otherwise, returns (assignment array,type array,type distribution array). Assignment array takes the form such that element (i,j) gives the probability that type i is assigned to each object j.

Is parametrized such that the type distribution represents unit capacity. Because prices enter the price mechanism via an incentive constraint, this solver does not support inputs for which price_type!="none" and info_bool==false.

Let us consider an example where the input is constructed using gen.ProblemGeneratorSimple.

julia> using DataFrames, Distributions

julia> include(abspath("../../../../basic-program-share/src/solver/modules/gen/src/gen.jl"))
Main.##ex-#286.gen

julia> cap_df = DataFrame(cap =[.3,.7,.5]); util_df = DataFrame(mean =[10,15,20]); shock_df = DataFrame(shock=[.1,.9])
2×1 DataFrames.DataFrame
│ Row │ shock   │
│     │ Float64 │
├─────┼─────────┤
│ 1   │ 0.1     │
│ 2   │ 0.9     │

julia> input = gen.ProblemGeneratorSimple(util_df,shock_df,cap_df,Logistic(),"DF")
([7.802775422663781 12.802775422663782 17.80277542266378; 7.802775422663781 12.802775422663782 22.19722457733622; … ; 12.19722457733622 17.19722457733622 17.80277542266378; 12.19722457733622 17.19722457733622 22.19722457733622], [0.002876319846178732, 0.017348816213405102, 0.017348816213405102, 0.10464115261950774, 0.017348816213405102, 0.10464115261950774, 0.10464115261950775, 0.6311537736550827], [0.3 0.7 0.5])

julia> include(abspath("../../../../basic-program-share/src/solver/modules/solve/src/solve.jl"))
Main.##ex-#286.solve

julia> using JuMP, GLPK, MathOptInterface
WARNING: using JuMP.solve in module ##ex-#286 conflicts with an existing identifier.

julia> const MOI = MathOptInterface
MathOptInterface

julia> import JuMP: GenericAffExpr

julia> a,b,c,d = solve.SolverL(input[1],input[2],input[3])
ERROR: UndefVarError: price_type not defined

julia> welfare = a
ERROR: UndefVarError: a not defined

julia> assignments = b
ERROR: UndefVarError: b not defined

julia> corresponding_types = c
ERROR: UndefVarError: c not defined

julia> distribution_of_corresponding_types = d
ERROR: UndefVarError: d not defined

Next, let's consider an example where the input is given directly.

julia> using DataFrames, Distributions

julia> include(abspath("../../../../basic-program-share/src/solver/modules/solve/src/solve.jl"))
Main.##ex-#503.solve

julia> using JuMP, GLPK, MathOptInterface, Random
WARNING: using JuMP.solve in module ##ex-#503 conflicts with an existing identifier.

julia> const MOI = MathOptInterface
MathOptInterface

julia> import JuMP: GenericAffExpr

julia> rng = MersenneTwister(1234);

julia> type_arr = randn!(rng,zeros(4,2)); type_arr = broadcast(+,1,type_arr)
4×2 Array{Float64,2}:
 1.86735    1.8644
 0.0982562  3.21188
 0.505521   1.53281
 0.0970857  0.728265

julia> type_probs = randn!(rng,zeros(4,2)); type_probs = broadcast(+,1,type_probs); type_probs = broadcast(/,type_probs,sum(type_probs))
4×2 Array{Float64,2}:
 0.173305   0.13013
 0.0557193  0.329089
 0.0506993  0.0198687
 0.113132   0.128057

julia> cap_arr = Array{Float64}(undef, 2); cap_arr[1] = .5; cap_arr[2] = .8
0.8

julia> a,b,c,d = solve.SolverL(type_arr,type_probs,cap_arr)
ERROR: UndefVarError: price_type not defined

julia> welfare = a
ERROR: UndefVarError: a not defined

julia> assignments = b
ERROR: UndefVarError: b not defined

julia> corresponding_types = c
ERROR: UndefVarError: c not defined

julia> distribution_of_corresponding_types = d
ERROR: UndefVarError: d not defined

solve.SolverNL

solve.SolverNL — Method

SolverNL(type_arr,type_probs,cap_arr;infocon_bool=true,eff_bool=false,price_type="lower")

Applies ModelSolver with Ipopt Optimizer. Use for Nonlinear Programming (with prices)

This solver should be used for non-linear programming problems (typically using prices), as the Ipopt optimizer used is not optimized for linear programming and will underperform the optimizer used in SolverL.

Given the type array, type distribution, and capacity array, finds the optimal allocation given the following constraints:

Capacity: $\Sigma_{\theta \in \Theta} f_{\theta} m_{\theta a} \leq c_a, \forall a \in A$

Feasibility: $\Sigma_{a \in A} m_{\theta a} \leq 1, \forall \theta \in \Theta$

Non-negativity: $m_{\theta a} \geq 0, \forall (\theta,a) \in \Theta \times A$

(Incentive Compatibility, applied when infocon_bool==1): $\Sigma_{a \in A} u_{\theta a} m_{\theta a} - \Sigma_{a \in A} u_{\theta a} m_{\theta' a} \geq 0, \forall \theta,\theta' \in \Theta$

(Efficiency, applied when eff_bool==1): pending

Saves the following tuple when print_bool ==1 to the folder given by folder ==path/to/folder: (assignment array,type array,type distribution array, incentive constraint designation). Otherwise, returns (assignment array,type array,type distribution array). Assignment array takes the form such that element (i,j) gives the probability that type i is assigned to each object j.

Is parametrized such that the type distribution represents unit capacity. Because prices enter the price mechanism via an incentive constraint, this solver does not support inputs for which price_type!="none" and info_bool==false.

Let us consider an example of a lower-bound price-regime where the input is constructed using gen.ProblemGeneratorSimple.

julia> using DataFrames, Distributions

julia> include(abspath("../../../../basic-program-share/src/solver/modules/gen/src/gen.jl"))
Main.##ex-#720.gen

julia> cap_df = DataFrame(cap =[.3,.7]); util_df = DataFrame(mean =[10,15]); shock_df = DataFrame(shock=[.1,.9])
2×1 DataFrames.DataFrame
│ Row │ shock   │
│     │ Float64 │
├─────┼─────────┤
│ 1   │ 0.1     │
│ 2   │ 0.9     │

julia> input = gen.ProblemGeneratorSimple(util_df,shock_df,cap_df,Logistic(),"DF")
([7.802775422663781 12.802775422663782; 7.802775422663781 17.19722457733622; 12.19722457733622 12.802775422663782; 12.19722457733622 17.19722457733622], [0.020225136059583836, 0.12198996883291284, 0.12198996883291284, 0.7357949262745904], [0.3 0.7])

julia> include(abspath("../../../../basic-program-share/src/solver/modules/solve/src/solve.jl"))
Main.##ex-#720.solve

julia> using JuMP, GLPK, MathOptInterface, Ipopt
WARNING: using JuMP.solve in module ##ex-#720 conflicts with an existing identifier.

julia> const MOI = MathOptInterface
MathOptInterface

julia> import JuMP: GenericAffExpr

julia> a,b,c,d,f = solve.SolverNL(input[1],input[2],input[3],price_type="lower")

******************************************************************************
This program contains Ipopt, a library for large-scale nonlinear optimization.
 Ipopt is released as open source code under the Eclipse Public License (EPL).
         For more information visit http://projects.coin-or.org/Ipopt
******************************************************************************

(15.606285932482994, [0.4044400892041755 0.5856717892364642; 0.0 1.0000000106509994; 1.0000000106509994 0.0; 0.2306759374157832 0.7693240718092164], [7.802775422663781 12.802775422663782; 7.802775422663781 17.19722457733622; 12.19722457733622 12.802775422663782; 12.19722457733622 17.19722457733622], [0.020225136059583836, 0.12198996883291284, 0.12198996883291284, 0.7357949262745904], [173.49724040838936, 66.91258873628551])

julia> welfare = a
15.606285932482994

julia> assignments = b
4×2 Array{Float64,2}:
 0.40444   0.585672
 0.0       1.0
 1.0       0.0
 0.230676  0.769324

julia> corresponding_types = c
4×2 Array{Float64,2}:
  7.80278  12.8028
  7.80278  17.1972
 12.1972   12.8028
 12.1972   17.1972

julia> distribution_of_corresponding_types = d
4-element Array{Float64,1}:
 0.020225136059583836
 0.12198996883291284
 0.12198996883291284
 0.7357949262745904

julia> prices = f
2-element Array{Float64,1}:
 173.49724040838936
  66.91258873628551

Now, let us consider an example of an upper-bound price-regime where the input is constructed using gen.ProblemGeneratorSimple.

julia> using DataFrames, Distributions

julia> include(abspath("../../../../basic-program-share/src/solver/modules/gen/src/gen.jl"))
Main.##ex-#937.gen

julia> cap_df = DataFrame(cap =[.3,.7]); util_df = DataFrame(mean =[10,15]); shock_df = DataFrame(shock=[.1,.9])
2×1 DataFrames.DataFrame
│ Row │ shock   │
│     │ Float64 │
├─────┼─────────┤
│ 1   │ 0.1     │
│ 2   │ 0.9     │

julia> input = gen.ProblemGeneratorSimple(util_df,shock_df,cap_df,Logistic(),"DF")
([7.802775422663781 12.802775422663782; 7.802775422663781 17.19722457733622; 12.19722457733622 12.802775422663782; 12.19722457733622 17.19722457733622], [0.020225136059583836, 0.12198996883291284, 0.12198996883291284, 0.7357949262745904], [0.3 0.7])

julia> include(abspath("../../../../basic-program-share/src/solver/modules/solve/src/solve.jl"))
Main.##ex-#937.solve

julia> using JuMP, GLPK, MathOptInterface, Ipopt
WARNING: using JuMP.solve in module ##ex-#937 conflicts with an existing identifier.

julia> const MOI = MathOptInterface
MathOptInterface

julia> import JuMP: GenericAffExpr

julia> a,b,c,d,f = solve.SolverNL(input[1],input[2],input[3],price_type="upper")
(15.606285932483049, [0.484755883124659 0.5053559953154554; 0.0 1.0000000106510303; 1.0000000106510303 0.0; 0.2284682599750009 0.7715317492500038], [7.802775422663781 12.802775422663782; 7.802775422663781 17.19722457733622; 12.19722457733622 12.802775422663782; 12.19722457733622 17.19722457733622], [0.020225136059583836, 0.12198996883291284, 0.12198996883291284, 0.7357949262745904], [44515.285041346884, 98839.49336303609])

julia> welfare = a
15.606285932483049

julia> assignments = b
4×2 Array{Float64,2}:
 0.484756  0.505356
 0.0       1.0
 1.0       0.0
 0.228468  0.771532

julia> corresponding_types = c
4×2 Array{Float64,2}:
  7.80278  12.8028
  7.80278  17.1972
 12.1972   12.8028
 12.1972   17.1972

julia> distribution_of_corresponding_types = d
4-element Array{Float64,1}:
 0.020225136059583836
 0.12198996883291284
 0.12198996883291284
 0.7357949262745904

julia> prices = f
2-element Array{Float64,1}:
 44515.285041346884
 98839.49336303609

Finally, let us consider an example of a raffle regime where the input is constructed using gen.ProblemGeneratorSimple.

julia> using DataFrames, Distributions

julia> include(abspath("../../../../basic-program-share/src/solver/modules/gen/src/gen.jl"))
Main.##ex-#1154.gen

julia> cap_df = DataFrame(cap =[.3,.7]); util_df = DataFrame(mean =[10,15]); shock_df = DataFrame(shock=[.1,.9])
2×1 DataFrames.DataFrame
│ Row │ shock   │
│     │ Float64 │
├─────┼─────────┤
│ 1   │ 0.1     │
│ 2   │ 0.9     │

julia> input = gen.ProblemGeneratorSimple(util_df,shock_df,cap_df,Logistic(),"DF")
([7.802775422663781 12.802775422663782; 7.802775422663781 17.19722457733622; 12.19722457733622 12.802775422663782; 12.19722457733622 17.19722457733622], [0.020225136059583836, 0.12198996883291284, 0.12198996883291284, 0.7357949262745904], [0.3 0.7])

julia> include(abspath("../../../../basic-program-share/src/solver/modules/solve/src/solve.jl"))
Main.##ex-#1154.solve

julia> using JuMP, GLPK, MathOptInterface, Ipopt
WARNING: using JuMP.solve in module ##ex-#1154 conflicts with an existing identifier.

julia> const MOI = MathOptInterface
MathOptInterface

julia> import JuMP: GenericAffExpr

julia> a,b,c,d,f = solve.SolverNL(input[1],input[2],input[3],price_type="raffle")
(15.544663959147696, [0.16268977944423554 0.8373094728484457; 0.1133097730243459 0.8850509239896506; 0.9999998495992996 1.4840178097304194e-7; 0.21853503157044468 0.7814649689814074], [7.802775422663781 12.802775422663782; 7.802775422663781 17.19722457733622; 12.19722457733622 12.802775422663782; 12.19722457733622 17.19722457733622], [0.020225136059583836, 0.12198996883291284, 0.12198996883291284, 0.7357949262745904], [1.000000004639284, 1.0678488537878534])

julia> welfare = a
15.544663959147696

julia> assignments = b
4×2 Array{Float64,2}:
 0.16269   0.837309
 0.11331   0.885051
 1.0       1.48402e-7
 0.218535  0.781465

julia> corresponding_types = c
4×2 Array{Float64,2}:
  7.80278  12.8028
  7.80278  17.1972
 12.1972   12.8028
 12.1972   17.1972

julia> distribution_of_corresponding_types = d
4-element Array{Float64,1}:
 0.020225136059583836
 0.12198996883291284
 0.12198996883291284
 0.7357949262745904

julia> prices = f
2-element Array{Float64,1}:
 1.000000004639284
 1.0678488537878534

As we can see, here the bounds are rather tight, and the proposition that the raffle regime will obtain a welfare lower than that of the optimal price regime is corroborated.