solve.jl Documentation

solve.SolverGLPK

solve.SolverGLPK — Method

SolverGLPK(num_types,num_objects,type_arr,type_probs,cap_arr,folder,print_bool,infocon_bool,eff_bool)

Applies ModelSolver with GLPK Optimizer.

Let us consider an example where the input is constructed using gen.ProblemGeneratorSimple.

julia> using DataFrames, Distributions

julia> include(abspath("../../../../basic-program-share/src/solver/modules/gen/src/gen.jl"))
Main.##ex-#255.gen

julia> cap_df = DataFrame(cap =[.3,.7,.5]); util_df = DataFrame(mean =[10,15,20]); shock_df = DataFrame(shock=[.1,.9])
2×1 DataFrames.DataFrame
│ Row │ shock   │
│     │ Float64 │
├─────┼─────────┤
│ 1   │ 0.1     │
│ 2   │ 0.9     │

julia> input = gen.ProblemGeneratorSimple(util_df,shock_df,cap_df,Logistic(),"DF")
([7.802775422663781 12.802775422663782 17.80277542266378; 7.802775422663781 12.802775422663782 22.19722457733622; … ; 12.19722457733622 17.19722457733622 17.80277542266378; 12.19722457733622 17.19722457733622 22.19722457733622], [0.002876319846178732, 0.017348816213405102, 0.017348816213405102, 0.10464115261950774, 0.017348816213405102, 0.10464115261950774, 0.10464115261950775, 0.6311537736550827], [0.3 0.7 0.5])

julia> include(abspath("../../../../basic-program-share/src/solver/modules/solve/src/solve.jl"))
Main.##ex-#255.solve

julia> using JuMP, GLPK, MathOptInterface
WARNING: using JuMP.solve in module ##ex-#255 conflicts with an existing identifier.

julia> const MOI = MathOptInterface
MathOptInterface

julia> import JuMP: GenericAffExpr

julia> assignments = solve.SolverGLPK(input[1],input[2],input[3],"output",0,1,0)[1]
8×3 Array{Float64,2}:
 0.0657672  0.0       0.846114
 0.0657672  0.0       0.846114
 0.0        1.0       0.0
 0.0        0.505925  0.477269
 0.261366   0.0       0.738634
 0.261366   0.0       0.738634
 0.0        1.0       0.0
 0.0657672  0.391197  0.543036

julia> corresponding_types = solve.SolverGLPK(input[1],input[2],input[3],"output",0,1,0)[2]
8×3 Array{Float64,2}:
  7.80278  12.8028  17.8028
  7.80278  12.8028  22.1972
  7.80278  17.1972  17.8028
  7.80278  17.1972  22.1972
 12.1972   12.8028  17.8028
 12.1972   12.8028  22.1972
 12.1972   17.1972  17.8028
 12.1972   17.1972  22.1972

julia> distribution_of_corresponding_types = solve.SolverGLPK(input[1],input[2],input[3],"output",0,1,0)[3]
8-element Array{Float64,1}:
 0.002876319846178732
 0.017348816213405102
 0.017348816213405102
 0.10464115261950774
 0.017348816213405102
 0.10464115261950774
 0.10464115261950775
 0.6311537736550827

Next, let consider an example where the input is given directly.

julia> using DataFrames, Distributions

julia> include(abspath("../../../../basic-program-share/src/solver/modules/solve/src/solve.jl"))
Main.##ex-#353.solve

julia> using JuMP, GLPK, MathOptInterface, Random
WARNING: using JuMP.solve in module ##ex-#353 conflicts with an existing identifier.

julia> const MOI = MathOptInterface
MathOptInterface

julia> import JuMP: GenericAffExpr

julia> rng = MersenneTwister(1234);

julia> type_arr = randn!(rng,zeros(4,2)); type_arr = broadcast(+,1,type_arr)
4×2 Array{Float64,2}:
 1.86735    1.8644
 0.0982562  3.21188
 0.505521   1.53281
 0.0970857  0.728265

julia> type_probs = randn!(rng,zeros(4,2)); type_probs = broadcast(+,1,type_probs); type_probs = broadcast(/,type_probs,sum(type_probs))
4×2 Array{Float64,2}:
 0.173305   0.13013
 0.0557193  0.329089
 0.0506993  0.0198687
 0.113132   0.128057

julia> cap_arr = Array{Float64}(undef, 2); cap_arr[1] = .5; cap_arr[2] = .8
0.8

julia> assignments = solve.SolverGLPK(type_arr,type_probs,cap_arr,"output",0,1,0)[1]
4×2 Array{Float64,2}:
 1.0  0.0
 0.0  1.0
 0.0  1.0
 0.0  1.0

julia> corresponding_types = solve.SolverGLPK(type_arr,type_probs,cap_arr,"output",0,1,0)[2]
4×2 Array{Float64,2}:
 1.86735    1.8644
 0.0982562  3.21188
 0.505521   1.53281
 0.0970857  0.728265

julia> distribution_of_corresponding_types = solve.SolverGLPK(type_arr,type_probs,cap_arr,"output",0,1,0)[3]
4×2 Array{Float64,2}:
 0.173305   0.13013
 0.0557193  0.329089
 0.0506993  0.0198687
 0.113132   0.128057

solve.ModelSolver

solve.ModelSolver — Method

ModelSolver(num_types,num_objects,type_arr,type_probs,cap_arr,m,folder,print_bool,infocon_bool,eff_bool)

Given a model m, ModelSolver solves the optimization problem subject to the appropriate constrains.

Given the number of types, number of means, capacity array, type arrays and type distribution, finds the optimal allocation given the following constraints:

Capacity: $\Sigma_{\theta \in \Theta} f_{\theta} m_{\theta a} \leq c_a, \forall a \in A$

Feasibility: $\Sigma_{a \in A} m_{\theta a} \leq 1, \forall \theta \in \Theta$

Non-negativity: $m_{\theta a} \geq 0, \forall (\theta,a) \in \Theta \times A$

(Incentive Compatibility, applied when infocon_bool==1): $\Sigma_{a \in A} u_{\theta a} m_{\theta a} - \Sigma_{a \in A} u_{\theta a} m_{\theta' a} \geq 0, \forall \theta,\theta' \in \Theta$

(Efficiency, applied when eff_bool==1): pending

Saves the following tuple when print_bool ==1: (assignment array,type array,type distribution array, vector of type vectors by shock, incentive constraint designation). Otherwise, returns (assignment array,type array,type distribution array, vector of type vectors by shock). Assignment array takes the form such that element (i,j) gives the probability that type i is assigned to each object j.

Is parametrized such that the type distribution represents unit capacity.